+++ src/App/Fossa/Container/Scan.hs
@@ -11,7 +11,6 @@ module App.Fossa.Container.Scan (
 ) where
 
 import App.Fossa.Config.Analyze (UseGitBackedCargoLocators, WithoutDefaultFilters)
-import App.Fossa.Config.Container.Analyze (GoBinaryDiscovery)
 import App.Fossa.Config.Container.Common (ImageText (unImageText))
 import App.Fossa.Container.Sources.DockerArchive (analyzeFromDockerArchive, revisionFromDockerArchive)
 import App.Fossa.Container.Sources.DockerEngine (analyzeFromDockerEngine, revisionFromDockerEngine)
@@ -91,12 +90,11 @@ scanImage ::
   AllFilters ->
   Flag WithoutDefaultFilters ->
   Bool ->
-  Flag GoBinaryDiscovery ->
   ImageText ->
   Text ->
   Text ->
   m ContainerScan
-scanImage useGitBackedCargo filters withoutDefaultFilters systemDepsOnly goBinaryDiscovery imgText dockerHost imageArch = do
+scanImage useGitBackedCargo filters withoutDefaultFilters systemDepsOnly imgText dockerHost imageArch = do
   parsedSource <- runDockerEngineApi dockerHost $ parseContainerImageSource (unImageText imgText) imageArch
   circePoweredScan <- withSystemTempDir "fossa-container-export-tmp" $ \dir -> do
     tarball <- runWithCirceReexport imgText dir
@@ -104,19 +102,19 @@ scanImage useGitBackedCargo filters withoutDefaultFilters systemDepsOnly goBinar
     traverse (correctCirceSource parsedSource) scan
   maybe (legacyScan parsedSource) pure circePoweredScan
   where
-    analyzeTarball = context "Analyzing docker archive" . analyzeFromDockerArchive useGitBackedCargo systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters
+    analyzeTarball = context "Analyzing docker archive" . analyzeFromDockerArchive useGitBackedCargo systemDepsOnly filters withoutDefaultFilters
     legacyScan src = do
       case src of
         DockerArchive tarball -> context "Analyzing tarball" $ analyzeTarball tarball
         DockerEngine imgTag ->
           context "Analyzing via Docker engine API" $
-            analyzeFromDockerEngine useGitBackedCargo systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters dockerHost imgTag
+            analyzeFromDockerEngine useGitBackedCargo systemDepsOnly filters withoutDefaultFilters dockerHost imgTag
         Podman img ->
           context "Analyzing via podman" $
-            analyzeFromPodman useGitBackedCargo systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters img
+            analyzeFromPodman useGitBackedCargo systemDepsOnly filters withoutDefaultFilters img
         Registry registrySrc ->
           context "Analyzing via registry" $
-            analyzeFromRegistry useGitBackedCargo systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters registrySrc
+            analyzeFromRegistry useGitBackedCargo systemDepsOnly filters withoutDefaultFilters registrySrc
 
 scanImageNoAnalysis ::
   ( Has Diagnostics sig m