+++ src/App/Fossa/Container/Sources/DockerEngine.hs
@@ -7,7 +7,6 @@ module App.Fossa.Container.Sources.DockerEngine (
 ) where
 
 import App.Fossa.Config.Analyze (UseGitBackedCargoLocators, WithoutDefaultFilters)
-import App.Fossa.Config.Container.Analyze (GoBinaryDiscovery)
 import App.Fossa.Container.Sources.DockerArchive (analyzeFromDockerArchive, listTargetsFromDockerArchive, revisionFromDockerArchive)
 import Container.Types (ContainerScan)
 import Control.Carrier.DockerEngineApi (runDockerEngineApi)
@@ -57,17 +56,16 @@ analyzeFromDockerEngine ::
   ) =>
   UseGitBackedCargoLocators ->
   Bool ->
-  Flag GoBinaryDiscovery ->
   AllFilters ->
   Flag WithoutDefaultFilters ->
   Text ->
   Text ->
   m ContainerScan
-analyzeFromDockerEngine useGitBackedCargo systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters engineHost imgTag =
+analyzeFromDockerEngine useGitBackedCargo systemDepsOnly filters withoutDefaultFilters engineHost imgTag =
   runFromDockerEngine
     engineHost
     imgTag
-    $ analyzeFromDockerArchive useGitBackedCargo systemDepsOnly goBinaryDiscovery filters withoutDefaultFilters
+    $ analyzeFromDockerArchive useGitBackedCargo systemDepsOnly filters withoutDefaultFilters
 
 listTargetsFromDockerEngine ::
   ( Has Diagnostics sig m